requirements.txt
streamlit==1.35.0
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.6
faiss-cpu==1.8.0
pymupdf==1.24.5
sentence-transformers==3.0.1
python-dotenv==1.0.1
tiktoken==0.7.0
